+++ CHANGELOG.md
@@ -3,7 +3,7 @@
     Placeholder for the next version (at the beginning of the line):
     ## **WORK IN PROGRESS**
 -->
-## **WORK IN PROGRESS**
+## 3.7.0 (2023-11-29)
 * Added `-lf` option to run the lock file update with the `--force` flag
 * Dependency upgrades